+++ jest.config.js
@@ -1,7 +1,7 @@
 module.exports = {
   preset: 'ts-jest',
   testMatch: ['**/*.spec.[jt]s?(x)'],
-  testTimeout: process.env.CI ? 25000 : 10000,
+  testTimeout: process.env.CI ? 30000 : 10000,
   globalSetup: './scripts/jestGlobalSetup.js',
   globalTeardown: './scripts/jestGlobalTeardown.js',
   testEnvironment: './scripts/jestEnv.js',