+++ packages/playground/optimize-deps/__tests__/optimize-deps.spec.ts
@@ -30,10 +30,18 @@ test('forced include', async () => {
   expect(await page.textContent('.force-include')).toMatch(`[success]`)
 })
 
+test('import * from optimized dep', async () => {
+  expect(await page.textContent('.import-star')).toMatch(`[success]`)
+})
+
 test('dep with css import', async () => {
   expect(await getColor('h1')).toBe('red')
 })
 
 test('dep w/ non-js files handled via plugin', async () => {
   expect(await page.textContent('.plugin')).toMatch(`[success]`)
+})
+
+test('vue + vuex', async () => {
+  expect(await page.textContent('.vue')).toMatch(`[success]`)
 })